// order-service/src/types/proto/orders.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.10.1
//   protoc               v6.33.1
// source: proto/orders.proto

/* eslint-disable */
import { GrpcMethod, GrpcStreamMethod } from "@nestjs/microservices";
import { Observable } from "rxjs";

export const protobufPackage = "orders";

export interface CreateOrderReq {
  productId: number;
  quantity: number;
  unitPrice: number;
}

export interface CreateOrderListReq {
  details: CreateOrderReq[];
}

export interface GetOrderIDReq {
  orderId: number;
}

/** Empty request for getting all orders */
export interface GetOrdersReq {
}

export interface OrderRes {
  orderId: number;
  details: OrderDetail[];
  totalAmount: number;
  createdAt: string;
}

export interface OrderDetail {
  productId: number;
  quantity: number;
  unitPrice: number;
  subTotal: number;
}

export interface OrderListRes {
  orders: OrderRes[];
}

export interface DeleteOrderRes {
  success: boolean;
  message: string;
}

export const ORDERS_PACKAGE_NAME = "orders";

export interface OrderServiceClient {
  createOrder(request: CreateOrderListReq): Observable<OrderRes>;

  getOrder(request: GetOrderIDReq): Observable<OrderRes>;

  getOrders(request: GetOrdersReq): Observable<OrderListRes>;

  deleteOrder(request: GetOrderIDReq): Observable<DeleteOrderRes>;
}

export interface OrderServiceController {
  createOrder(request: CreateOrderListReq): Promise<OrderRes> | Observable<OrderRes> | OrderRes;

  getOrder(request: GetOrderIDReq): Promise<OrderRes> | Observable<OrderRes> | OrderRes;

  getOrders(request: GetOrdersReq): Promise<OrderListRes> | Observable<OrderListRes> | OrderListRes;

  deleteOrder(request: GetOrderIDReq): Promise<DeleteOrderRes> | Observable<DeleteOrderRes> | DeleteOrderRes;
}

export function OrderServiceControllerMethods() {
  return function (constructor: Function) {
    const grpcMethods: string[] = ["createOrder", "getOrder", "getOrders", "deleteOrder"];
    for (const method of grpcMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcMethod("OrderService", method)(constructor.prototype[method], method, descriptor);
    }
    const grpcStreamMethods: string[] = [];
    for (const method of grpcStreamMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcStreamMethod("OrderService", method)(constructor.prototype[method], method, descriptor);
    }
  };
}

export const ORDER_SERVICE_NAME = "OrderService";